+++ NEWS.md
@@ -1,5 +1,33 @@
 <!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
 
+# RMariaDB 1.2.1 (2021-12-20)
+
+## Features
+
+- Upgrade to mariadb-connector-c 3.2.5 on Windows, with built-in support for the `caching_sha2_password` plugin (#134, #248, @jeroen).
+
+## Internal
+
+- Make method definition more similar to S3. All `setMethod()` calls refer to top-level functions (#250).
+
+
+# RMariaDB 1.2.0 (2021-12-12)
+
+## Features
+
+- BLOBs are returned as `blob::blob()` objects (#126, #243).
+- `dbWriteTable()` and `dbAppendTable()` are much faster thanks to `LOAD DATA LOCAL INFILE`. To activate this, `load_data_local_infile = TRUE` must be passed to `dbConnect()`. The readr package is required (#11, #223).
+
+## Bug fixes
+
+- Writing dates prior to 1970 no longer crashes on Windows (#232, #245).
+
+## Internal
+
+- Add test for reading and writing JSON data type as string (#127, #246).
+- Update for compatibility with DBItest 1.7.2 (#228).
+
+
 # RMariaDB 1.1.2 (2021-09-06)
 
 ## Licensing